Address code review feedback: remove unnecessary cast and fix format specifier
Agent-Logs-Url: https://github.com/posixninja/libmacho/sessions/952e110e-908e-465c-8d1e-5079760dc5f1

+++ src/symtab.c
@@ -76,7 +76,7 @@ macho_symtab_t* macho_symtab_load(unsigned char* cmd, unsigned char* data, uint8
 				uint32_t strx = syms[i].n_strx;
 				symtab->symbols[i].n_type  = syms[i].n_type;
 				symtab->symbols[i].n_sect  = syms[i].n_sect;
-				symtab->symbols[i].n_desc  = (int16_t)syms[i].n_desc;
+				symtab->symbols[i].n_desc  = syms[i].n_desc;
 				symtab->symbols[i].n_value = syms[i].n_value;
 				if (strx >= symtab->cmd->strsize) {
 					symtab->symbols[i].n_un.n_name = NULL;
